Added Roles

+++ src/main/java/com/musicplanet/dalinterfaces/IUserDal.java
@@ -5,10 +5,10 @@ import java.util.List;
 import java.util.Optional;
 
 public interface IUserDal {
-    Optional<User> getById (int id);
+    Optional<User> getById (Long id);
     List<User> getAll();
     void save(User user);
-    void delete(int id);
+    void delete(Long id);
 
     Optional<User> getByEmail(String email);
     Optional<User> getByUsername(String username);
