+++ src/main/java/com/musicplanet/repository/User/IUserRepository.java
@@ -16,7 +16,7 @@ import java.util.Optional;
 public interface IUserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
     Optional<User> findByUsername(String username);
-    Optional<User> findById(int id);
+    Optional<User> findById(Long id);
 
     @Transactional
     @Modifying
